django-tdlogbook/requirements.txt
Django==4.2
djangorestframework==3.14.0
django-cors-headers>=4.0.0
psycopg2-binary>=2.9.10
celery==5.3.4
redis==5.0.1
python-dateutil==2.8.2
pytz==2023.3
